+++ transact/src/main/java/dev/dbos/transact/database/MetricData.java
@@ -1,3 +1,3 @@
 package dev.dbos.transact.database;
 
-public record MetricData(String metricType, String metricName, int value) {}
+public record MetricData(String metricType, String metricName, long value) {}
